Implement `NavigationPaneWidget`
